requirements.txt
blinker==1.6.2 ; python_full_version == "3.11.4"
boto3==1.28.39 ; python_full_version == "3.11.4"
botocore==1.31.39 ; python_full_version == "3.11.4"
certifi==2023.7.22 ; python_full_version == "3.11.4"
charset-normalizer==3.2.0 ; python_full_version == "3.11.4"
click==8.1.7 ; python_full_version == "3.11.4"
colorama==0.4.6 ; python_full_version == "3.11.4" and platform_system == "Windows"
contourpy==1.1.0 ; python_full_version == "3.11.4"
cycler==0.11.0 ; python_full_version == "3.11.4"
flask==2.3.3 ; python_full_version == "3.11.4"
fonttools==4.42.1 ; python_full_version == "3.11.4"
idna==3.4 ; python_full_version == "3.11.4"
itsdangerous==2.1.2 ; python_full_version == "3.11.4"
jinja2==3.1.2 ; python_full_version == "3.11.4"
jmespath==1.0.1 ; python_full_version == "3.11.4"
kiwisolver==1.4.5 ; python_full_version == "3.11.4"
markupsafe==2.1.3 ; python_full_version == "3.11.4"
matplotlib==3.7.2 ; python_full_version == "3.11.4"
numpy==1.25.2 ; python_full_version == "3.11.4"
opencv-python==4.7.0.72 ; python_full_version == "3.11.4"
packaging==23.1 ; python_full_version == "3.11.4"
pendulum==2.1.2 ; python_full_version == "3.11.4"
pillow==10.0.0 ; python_full_version == "3.11.4"
pyparsing==3.0.9 ; python_full_version == "3.11.4"
python-dateutil==2.8.2 ; python_full_version == "3.11.4"
python-dotenv==1.0.0 ; python_full_version == "3.11.4"
pytzdata==2020.1 ; python_full_version == "3.11.4"
requests==2.31.0 ; python_full_version == "3.11.4"
s3transfer==0.6.2 ; python_full_version == "3.11.4"
six==1.16.0 ; python_full_version == "3.11.4"
urllib3==1.26.16 ; python_full_version == "3.11.4"
werkzeug==2.3.7 ; python_full_version == "3.11.4"
